Use `decimal` .Net type for the `Price` instead of `double`.

+++ Tsk.HttpApi/Products/ForAdmins/DataTransferObjects.cs
@@ -9,7 +9,7 @@ public class ProductDto
 {
     public required Guid Id { get; init; }
     public required string Title { get; init; }
-    public required double Price { get; init; }
+    public required decimal Price { get; init; }
     public required bool IsForSale { get; init; }
 }
 
@@ -22,7 +22,7 @@ public class CreateProductDto
 
     [Required]
     [GreaterThan(0)]
-    public required double Price { get; init; }
+    public required decimal Price { get; init; }
 }
 
 [PublicAPI]
@@ -34,5 +34,5 @@ public class UpdateProductDto
 
     [Required]
     [GreaterThan(0)]
-    public required double Price { get; init; }
+    public required decimal Price { get; init; }
 }
