+++ Tsk.HttpApi/Products/Model.cs
@@ -12,7 +12,7 @@ public class ProductEntity
     public required string Title { get; set; }
 
     [Column("price")]
-    public required double Price { get; set; }
+    public required decimal Price { get; set; }
 
     [Column("is_for_sale")]
     public required bool IsForSale { get; set; }
